+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add support for backslashed character literals (e.g., `\t`, `\\`, `\u{00e9}`, etc.)
 * Add support for `uninterp`reted spec functions (see [verus#1473](https://github.com/verus-lang/verus/pull/1473))
+* Add support for the `axiom` syntax (see [verus#1518](https://github.com/verus-lang/verus/pull/1518))
 
 # v0.5.3